SSpline.jl Documentation

Introduction

Construction

Spline Functions

SpectralVis.jl Documentation

SSpline.jl Documentation

About

Functions

Index

References

Constrained cubic splines

Timings

Linear Splines

Cubic Comparison

SSpline Interpolation timings

Spline creation timings

About

This package (SSpline = short for SpectralSpline) provides a collection of methods specifically tailored to deal with spectral calculations. It provides functions and methods for the creation of univariate splines of different degrees and types and various interpolation and extrapolation methods.

Functions

SSpline.Mbound — Method

Mbound(M,style...)

Returns vector M (moments) according to boundary conditions defined by style...

Helper function for cubicspline.

SSpline.clampedspline — Function

clampedspline(x::Vector, y::Vector, ds = 0.0, de = 0.0)

Equivalent to cubicspline(x::Vector, y::Vector, ds, de).

Returns a clamped cubic spline.

Default arguments: 1st and last derivatives = 0.0

SSpline.cubicspline — Method

cubicspline(x, y)

Fall back function, setting the default boundary condition to :natural

alias: naturalspline

SSpline.cubicspline — Method

cubicspline(x::Vector, y::Vector, style...)

alias: spline3, cspline

Creates a cubic spline. Boundary conditions are defined by style.

style...: settings for boundary conditions.

:natural: zero curvature at the start and end points of the spline.

aliases: cubicspline(x, y), naturalspline(x, y)

:clamped, ds, de: predefined 1st derivative of start (ds) and end points (de).

alias: clampedspline(x, y, ds, de)

:periodic: periodic boundary conditions. y values at start and end points have to be equal,

conditions at start and end points are identical. - alias: periodicspline(x, y)

:parabolic, :notaknot: parabolic runout spline, also known as not-a-knot spline

:cubic: cubic runout spline

:constrained: constrained cubic spline, see Kruger’s paper

:hermite: hermite spline

:kruger: constrained cubic spline, see Kruger’s paper. No overshooting.

:free: hermite spline with free derivative conditions. No overshooting.

:predef: predefined derivatives at knots

Examples:

julia> x=[0,1,2,3,4]; y=[-1,1,5,2,-3]; spline3(x,y,:natural)
CubicSpline([1.0, -3.0, 2.0, 0.0], [0.0, 3.0, -6.0, 0.0], [1.0, 4.0, 1.0, -5.0], [-1.0, 1.0, 5.0, 2.0])

SSpline.curvature — Method

curvature(spl::Spline,dλ)

Returns the 2nd derivative of the spline function (s'')

SSpline.curvrate — Method

curvrate(spl::Spline,dλ)

Returns the 3rd derivative of the spline function (s''').

SSpline.extrapolate — Function

extrapolate(spl, env = SpectralVis.SPECENV)

Extrapolate spline according to spectral environment.

SSpline.extrapolate — Method

extrapolate(spl, xrange, extr::Symbol)

Extrapolate a spline to the limits given by xrange, using the extrapolation type extr.

alias: extrap

Allowed modes for extr:

:zero

:boundary

:linear

:quadratic

:smooth (smooth blending to zero for Hermite splines)

SSpline.hermitespline — Function

hermitespline(x::Vector, y::Vector, style::Symbol = :kruger, args...)

Creates cubic hermite interpolation spline with defined 1st derivatives at all knots.

Allowed styles:

:kruger: Default value. Defines 1st derivatives using harmonic means to prevent overshoot at extremal knots.

:free: Simple mean values for 1st derivatives.

:predef: user defined 1st derivatives at all knots.

SSpline.hermitespline — Method

hermitespline(x::Vector, y::Vector, d::Vector) = hermitespline(x, y, :predef, d)

Equivalent to hermitespline(x, y, :predef, d)

SSpline.interpolate — Method

interp(spl::CubicSpline, xrange)

Returns tuple of vectors of spline values s at locations xs, as defined by the grid points at x+n*dλ (smoothed curve).

Examples:

for x=[0,10,30,50,70,90,100] and y=[30,130,150,150,170,220,320]

julia> cs = spline3(x,y,:natural)
CubicSpline([-0.015813397129186602, 0.009126794258373205, -0.0006937799043062205, -0.0013516746411483252, 0.007350478468899521, -0.01305023923444976], [0.0, -0.4744019138755981, 0.07320574162679427, 0.03157894736842105, -0.049521531100478466, 0.3915071770334928], [11.58133971291866, 6.83732057416268, -1.1866028708133973, 0.9090909090909092, 0.5502392344497609, 7.389952153110048], [30.0, 130.0, 150.0, 150.0, 170.0, 220.0], [0.0, 10.0, 30.0, 50.0, 70.0, 90.0, 100.0])

julia> interp(cs, 30.0:0.2:320.0)
([30.0, 30.2, 30.4, 30.6, 30.8, 31.0, 31.2, 31.4, 31.6, 31.8  …  318.2, 318.4, 318.6, 318.8, 319.0, 319.2, 319.4, 319.6, 319.8, 320.0], [150.00000000000003, 149.76560210526316, 149.53702736842106, 149.31424248803827, 149.09721416267942, 148.88590909090908, 148.68029397129186, 148.48033550239234, 148.28600038277511, 148.09725531100477  …  0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0])

SSpline.interpolate — Method

interpolate(spl::HermiteSpline, xrange)

Create interpolated vectors based on a Hermite spline

Examples:

for x=[0,10,30,50,70,90,100] and y=[30,130,150,150,170,220,320]

julia> hs = hermitespline(x,y,:kruger)
HermiteSpline([14.909090909090908, 0.18181818181818182, 0.0, 0.0, 0.5714285714285714, 0.16, 14.92], [30.0, 130.0, 150.0, 150.0, 170.0, 220.0, 320.0], [0.0, 10.0, 30.0, 50.0, 70.0, 90.0, 100.0])

julia> interp(hs, 30.0:0.2:320.0)
([30.0, 30.2, 30.4, 30.6, 30.8, 31.0, 31.2, 31.4, 31.6, 31.8  …  318.2, 318.4, 318.6, 318.8, 319.0, 319.2, 319.4, 319.6, 319.8, 320.0], [150.0, 150.0, 150.00000000000003, 149.99999999999997, 150.0, 150.0, 150.0, 150.0, 150.0, 150.0  …  0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0])

SSpline.interpolate — Method

interpolate(spl::LinearSpline, xrange)

Create interpolated vectors based on a linear spline

Examples:

for x=[0,10,30,50,70,90,100] and y=[30,130,150,150,170,220,320]

julia> ls = spline1(x,y)
LinearSpline([10.0, 1.0, 0.0, 1.0, 2.5, 10.0], [30.0, 130.0, 150.0, 150.0, 170.0, 220.0], [0.0, 10.0, 30.0, 50.0, 70.0, 90.0, 100.0])

julia> interp(ls, 30.0:0.2:320.0)
([30.0, 30.2, 30.4, 30.6, 30.8, 31.0, 31.2, 31.4, 31.6, 31.8  …  318.2, 318.4, 318.6, 318.8, 319.0, 319.2, 319.4, 319.6, 319.8, 320.0], [150.0, 150.0, 150.0, 150.0, 150.0, 150.0, 150.0, 150.0, 150.0, 150.0  …  0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0])

SSpline.interpolate — Method

interpolate(spl::NearestNeighborSpline, xrange)

Create interpolated vectors based on a nearest neighbor spline.

Examples:

for x=[0,10,30,50,70,90,100] and y=[30,130,150,150,170,220,320]

julia> nns = spline0(x,y)
NearestNeighborSpline([30.0, 130.0, 150.0, 150.0, 170.0, 220.0, 320.0], [0.0, 10.0, 30.0, 50.0, 70.0, 90.0, 100.0])

julia> interp(nns, 30.0:0.2:320.0)
([30.0, 30.2, 30.4, 30.6, 30.8, 31.0, 31.2, 31.4, 31.6, 31.8  …  318.2, 318.4, 318.6, 318.8, 319.0, 319.2, 319.4, 319.6, 319.8, 320.0], [150.0, 150.0, 150.0, 150.0, 150.0, 150.0, 150.0, 150.0, 150.0, 150.0  …  0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0])

SSpline.interpolate — Method

interpolate(spl::QuadraticSpline, xrange)

Create interpolated vectors based on a quadratic spline

Examples:

for x=[0,10,30,50,70,90,100] and y=[30,130,150,150,170,220,320]

julia> qs = spline2(x,y)
QuadraticSpline([1.0, -0.95, 0.9, -0.85, 0.925, -1.1, 0.0], [0.0, 20.0, -18.0, 18.0, -16.0, 21.0, -1.0], [30.0, 130.0, 150.0, 150.0, 170.0, 220.0, 0.0], [0.0, 10.0, 30.0, 50.0, 70.0, 90.0, 100.0])

julia> interp(qs, 30.0:0.2:320.0)
([30.0, 30.2, 30.4, 30.6, 30.8, 31.0, 31.2, 31.4, 31.6, 31.8  …  318.2, 318.4, 318.6, 318.8, 319.0, 319.2, 319.4, 319.6, 319.8, 320.0], [150.0, 146.436, 142.94400000000002, 139.52399999999997, 136.176, 132.9, 129.69600000000003, 126.56400000000002, 123.50399999999998, 120.51599999999999  …  2.2006465129835e-310, 2.2034094153224e-310, 2.2061740509752e-310, 2.20894041994186e-310, 2.2117085222224e-310, 2.21447835781683e-310, 2.2172499267252e-310, 2.2200232289474e-310, 2.22279826448355e-310, 2.22557503333353e-310])

SSpline.linearspline — Method

linearspline(x::Vector, y::Vector)

Linear spline interpolation.

aliases: spline1, lspline

SSpline.naturalspline — Method

naturalspline(x::Vector, y::Vector)

Equivalent to cubicspline(x::Vector, y::Vector, :natural).

SSpline.nearestneighbor — Method

nearestneighbor(x::Vector, y::Vector)

Nearest neighbor interpolation with the spline values in the center of intervals.

aliases: spline0, stepspline

SSpline.periodicspline — Method

periodicspline(x::Vector, y::Vector)

Equivalent to cubicspline(x::Vector, y::Vector, :periodic).

y values of start and end points have to be equal.

SSpline.quadraticspline — Function

quadraticspline(x::Vector, y::Vector, b1 = 0.0)

Quadratic spline interpolation.

aliases: spline2, qspline

Optional argument b1 sets the 1st derivative at the first knot. Default value = 0.0.

SSpline.rbound — Method

rbound(r,Δx,Δy,style...)

Returns vector r according to boundary conditions defined by style... Helper function for cubicspline.

Returns: r

SSpline.slope — Method

slope(spl::Spline,dλ)

Returns the pair of λ, s' (1st derivative of the spline function).

Index

SSpline.Mbound

SSpline.clampedspline

SSpline.clampedspline

SSpline.cubicspline

SSpline.cubicspline

SSpline.cubicspline

SSpline.curvature

SSpline.curvrate

SSpline.extrapolate

SSpline.extrapolate

SSpline.hermitespline

SSpline.hermitespline

SSpline.hermitespline

SSpline.interpolate

SSpline.interpolate

SSpline.interpolate

SSpline.interpolate

SSpline.interpolate

SSpline.linearspline

SSpline.linearspline

SSpline.naturalspline

SSpline.naturalspline

SSpline.nearestneighbor

SSpline.nearestneighbor

SSpline.periodicspline

SSpline.periodicspline

SSpline.quadraticspline

SSpline.quadraticspline

SSpline.rbound

SSpline.slope