Construction

Spline Functions

All available spline construction functions.

Given a pair of Vectors x and y, construct splines like follows:

Nearest Neighbor Splines

SSpline.nearestneighbor — Function

nearestneighbor(x::Vector, y::Vector)

Nearest neighbor interpolation with the spline values in the center of intervals.

aliases: spline0, stepspline

Linear Splines

SSpline.linearspline — Function

linearspline(x::Vector, y::Vector)

Linear spline interpolation.

aliases: spline1, lspline

Quadratic Splines

SSpline.quadraticspline — Function

quadraticspline(x::Vector, y::Vector, b1 = 0.0)

Quadratic spline interpolation.

aliases: spline2, qspline

Optional argument b1 sets the 1st derivative at the first knot. Default value = 0.0.

Cubic Splines

Hermite splines, defective cubic splines:

SSpline.hermitespline — Function

hermitespline(x::Vector, y::Vector, style::Symbol = :kruger, args...)

Creates cubic hermite interpolation spline with defined 1st derivatives at all knots.

Allowed styles:

:kruger: Default value. Defines 1st derivatives using harmonic means to prevent overshoot at extremal knots.

:free: Simple mean values for 1st derivatives.

:predef: user defined 1st derivatives at all knots.

hermitespline(x::Vector, y::Vector, d::Vector) = hermitespline(x, y, :predef, d)

Equivalent to hermitespline(x, y, :predef, d)

Cubic splines:

SSpline.cubicspline — Function

cubicspline(x::Vector, y::Vector, style...)

alias: spline3, cspline

Creates a cubic spline. Boundary conditions are defined by style.

style...: settings for boundary conditions.

:natural: zero curvature at the start and end points of the spline.

aliases: cubicspline(x, y), naturalspline(x, y)

:clamped, ds, de: predefined 1st derivative of start (ds) and end points (de).

alias: clampedspline(x, y, ds, de)

:periodic: periodic boundary conditions. y values at start and end points have to be equal,

conditions at start and end points are identical. - alias: periodicspline(x, y)

:parabolic, :notaknot: parabolic runout spline, also known as not-a-knot spline

:cubic: cubic runout spline

:constrained: constrained cubic spline, see Kruger’s paper

:hermite: hermite spline

:kruger: constrained cubic spline, see Kruger’s paper. No overshooting.

:free: hermite spline with free derivative conditions. No overshooting.

:predef: predefined derivatives at knots

Examples:

julia> x=[0,1,2,3,4]; y=[-1,1,5,2,-3]; spline3(x,y,:natural)
CubicSpline([1.0, -3.0, 2.0, 0.0], [0.0, 3.0, -6.0, 0.0], [1.0, 4.0, 1.0, -5.0], [-1.0, 1.0, 5.0, 2.0])

cubicspline(x, y)

Fall back function, setting the default boundary condition to :natural

alias: naturalspline

SSpline.naturalspline — Function

naturalspline(x::Vector, y::Vector)

Equivalent to cubicspline(x::Vector, y::Vector, :natural).

SSpline.clampedspline — Function

clampedspline(x::Vector, y::Vector, ds = 0.0, de = 0.0)

Equivalent to cubicspline(x::Vector, y::Vector, ds, de).

Returns a clamped cubic spline.

Default arguments: 1st and last derivatives = 0.0

SSpline.periodicspline — Function

periodicspline(x::Vector, y::Vector)

Equivalent to cubicspline(x::Vector, y::Vector, :periodic).

y values of start and end points have to be equal.